update to PR to rename manageCase to aac to avoid confusion with XUI manage case
